move racedetect logic to top level of src
