+++ src/CMakeLists.txt
@@ -41,8 +41,8 @@ set(racedetect-lib-sources
     Trace/EventImpl.cpp
     Trace/ProgramTrace.cpp
     Trace/ThreadTrace.cpp
-    RaceDetect/RaceDetect.cpp
-    Reporter/Reporter.cpp)
+    Reporter/Reporter.cpp
+    RaceDetect.cpp)
 add_library(racedetect STATIC ${racedetect-lib-sources})
 target_link_libraries(racedetect pta)
 target_include_directories(racedetect PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
