+++ src/RaceDetect.h
@@ -11,12 +11,7 @@ limitations under the License.
 
 #pragma once
 
-#include "Analysis/HappensBeforeGraph.h"
-#include "Analysis/LockSet.h"
-#include "Analysis/SharedMemory.h"
-#include "LanguageModel/RaceModel.h"
 #include "Reporter/Reporter.h"
-#include "Trace/ProgramTrace.h"
 
 namespace race {
